+++ scene2d/src/main/kotlin/ktx/scene2d/factory.kt
@@ -3,6 +3,9 @@ package ktx.scene2d
 import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.ui.*
 import com.badlogic.gdx.scenes.scene2d.ui.Tree.Node
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.InvocationKind
+import kotlin.contracts.contract
 import com.badlogic.gdx.utils.Array as GdxArray
 
 /**
@@ -28,12 +31,16 @@ inline fun <T : Actor> actor(actor: T, init: (@Scene2dDsl T).() -> Unit = {}): T
  * @return a new [Window] instance.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun RootWidget.window(
   title: String,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KWindow.() -> Unit = {}
-): KWindow = storeActor(KWindow(title, skin, style)).apply(init)
+): KWindow {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return storeActor(KWindow(title, skin, style)).apply(init)
+}
 
 /**
  * Constructs a top-level [Dialog] widget.
@@ -44,12 +51,16 @@ inline fun RootWidget.window(
  * @return a new [Dialog] instance.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun RootWidget.dialog(
   title: String,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KDialog.() -> Unit = {}
-): KDialog = storeActor(KDialog(title, skin, style)).apply(init)
+): KDialog {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return storeActor(KDialog(title, skin, style)).apply(init)
+}
 
 /**
  * Utility function for adding existing actors to the group with a type-safe builder init block.
@@ -62,7 +73,9 @@ inline fun RootWidget.dialog(
  * @param A actor type.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S, A : Actor> KWidget<S>.actor(actor: A, init: (@Scene2dDsl A).(S) -> Unit = {}): A {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
   actor.init(storeActor(actor))
   return actor
 }
@@ -76,11 +89,15 @@ inline fun <S, A : Actor> KWidget<S>.actor(actor: A, init: (@Scene2dDsl A).(S) -
  * @return a [Button] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.button(
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KButton.(S) -> Unit = {}
-): KButton = actor(KButton(skin, style), init)
+): KButton {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KButton(skin, style), init)
+}
 
 /**
  * @param minCheckedCount minimum amount of buttons checked at once.
@@ -93,12 +110,16 @@ inline fun <S> KWidget<S>.button(
  * [Button]-extending children of this widget will be added to the internal [ButtonGroup].
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.buttonGroup(
   minCheckedCount: Int,
   maxCheckedCount: Int,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KButtonTable.(S) -> Unit = {}
-): KButtonTable = actor(KButtonTable(minCheckedCount, maxCheckedCount, skin), init)
+): KButtonTable {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KButtonTable(minCheckedCount, maxCheckedCount, skin), init)
+}
 
 /**
  * @param text will be displayed as [CheckBox] text.
@@ -110,12 +131,16 @@ inline fun <S> KWidget<S>.buttonGroup(
  * @return a [CheckBox] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.checkBox(
   text: String,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KCheckBox.(S) -> Unit = {}
-): KCheckBox = actor(KCheckBox(text, skin, style), init)
+): KCheckBox {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KCheckBox(text, skin, style), init)
+}
 
 /**
  * @param init will be invoked with the widget as "this". Consumes actor container (usually a [Cell] or [Node]) that
@@ -124,9 +149,13 @@ inline fun <S> KWidget<S>.checkBox(
  * @return a [Container] instance added to this group. Note that this actor might store only a single child.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.container(
   init: KContainer<Actor>.(S) -> Unit = {}
-): KContainer<Actor> = actor(KContainer(), init)
+): KContainer<Actor> {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KContainer(), init)
+}
 
 /**
  * @param init will be invoked with the widget as "this". Consumes actor container (usually a [Cell] or [Node]) that
@@ -135,9 +164,13 @@ inline fun <S> KWidget<S>.container(
  * @return a [HorizontalGroup] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.horizontalGroup(
   init: KHorizontalGroup.(S) -> Unit = {}
-): KHorizontalGroup = actor(KHorizontalGroup(), init)
+): KHorizontalGroup {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KHorizontalGroup(), init)
+}
 
 /**
  * @param drawable name of a drawable stored in the chosen skin.
@@ -148,11 +181,15 @@ inline fun <S> KWidget<S>.horizontalGroup(
  * @return a [Image] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.image(
   drawable: String,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl Image).(S) -> Unit = {}
-): Image = actor(Image(skin.getDrawable(drawable)), init)
+): Image {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(Image(skin.getDrawable(drawable)), init)
+}
 
 /**
  * @param style name of the widget style. Defaults to [defaultStyle].
@@ -163,11 +200,15 @@ inline fun <S> KWidget<S>.image(
  * @return a [ImageButton] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.imageButton(
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KImageButton.(S) -> Unit = {}
-): KImageButton = actor(KImageButton(skin, style), init)
+): KImageButton {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KImageButton(skin, style), init)
+}
 
 /**
  * @param text will be displayed as [ImageTextButton] text.
@@ -179,12 +220,16 @@ inline fun <S> KWidget<S>.imageButton(
  * @return a [ImageTextButton] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.imageTextButton(
   text: String,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KImageTextButton.(S) -> Unit = {}
-): KImageTextButton = actor(KImageTextButton(text, skin, style), init)
+): KImageTextButton {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KImageTextButton(text, skin, style), init)
+}
 
 /**
  * @param text will be displayed on the label.
@@ -196,12 +241,16 @@ inline fun <S> KWidget<S>.imageTextButton(
  * @return a [Label] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.label(
   text: CharSequence,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl Label).(S) -> Unit = {}
-): Label = actor(Label(text, skin, style), init)
+): Label {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(Label(text, skin, style), init)
+}
 
 /**
  * @param items optional LibGDX array of list widget items. Defaults to null.
@@ -234,11 +283,13 @@ fun <I> KWidget<*>.listWidgetOf(
  * @param I type of items stored by this widget. Usually items are converted to string and displayed.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <I> KWidget<*>.listWidget(
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KListWidget<I>.() -> Unit = {}
 ): KListWidget<I> {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
   val list = KListWidget<I>(skin, style)
   storeActor(list)
   list.init()
@@ -260,6 +311,7 @@ inline fun <I> KWidget<*>.listWidget(
  * @return a [ProgressBar] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.progressBar(
   min: Float = 0f,
   max: Float = 1f,
@@ -268,7 +320,10 @@ inline fun <S> KWidget<S>.progressBar(
   style: String = if (vertical) defaultVerticalStyle else defaultHorizontalStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl ProgressBar).(S) -> Unit = {}
-): ProgressBar = actor(ProgressBar(min, max, step, vertical, skin, style), init)
+): ProgressBar {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(ProgressBar(min, max, step, vertical, skin, style), init)
+}
 
 /**
  * @param style name of the widget style. Defaults to [defaultStyle].
@@ -279,11 +334,15 @@ inline fun <S> KWidget<S>.progressBar(
  * @return a [ScrollPane] instance added to this group. Note that this actor may have only a single child.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.scrollPane(
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KScrollPane.(S) -> Unit = {}
-): KScrollPane = actor(KScrollPane(skin, style), init)
+): KScrollPane {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KScrollPane(skin, style), init)
+}
 
 /**
  * @param items optional LibGDX array of the [SelectBox] items. Defaults to null.
@@ -316,11 +375,13 @@ fun <I> KWidget<*>.selectBoxOf(
  * @param I type of items stored by this widget. Usually items are converted to string and displayed.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <I> KWidget<*>.selectBox(
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KSelectBox<I>.() -> Unit = {}
 ): KSelectBox<I> {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
   val selectBox = KSelectBox<I>(skin, style)
   storeActor(selectBox)
   selectBox.init()
@@ -342,6 +403,7 @@ inline fun <I> KWidget<*>.selectBox(
  * @return a [Slider] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.slider(
   min: Float = 0f,
   max: Float = 1f,
@@ -350,7 +412,10 @@ inline fun <S> KWidget<S>.slider(
   style: String = if (vertical) defaultVerticalStyle else defaultHorizontalStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl Slider).(S) -> Unit = {}
-): Slider = actor(Slider(min, max, step, vertical, skin, style), init)
+): Slider {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(Slider(min, max, step, vertical, skin, style), init)
+}
 
 /**
  * @param vertical true if the widget is vertical, false if horizontal.
@@ -363,12 +428,16 @@ inline fun <S> KWidget<S>.slider(
  * @return a [SplitPane] instance added to this group. Note that this actor can store only two children.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.splitPane(
   vertical: Boolean = false,
   style: String = if (vertical) defaultVerticalStyle else defaultHorizontalStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KSplitPane.(S) -> Unit = {}
-): KSplitPane = actor(KSplitPane(vertical, skin, style), init)
+): KSplitPane {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KSplitPane(vertical, skin, style), init)
+}
 
 /**
  * @param init will be invoked with the widget as "this". Consumes actor container (usually a [Cell] or [Node]) that
@@ -377,9 +446,13 @@ inline fun <S> KWidget<S>.splitPane(
  * @return a [Stack] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.stack(
   init: KStack.(S) -> Unit = {}
-): KStack = actor(KStack(), init)
+): KStack {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KStack(), init)
+}
 
 /**
  * @param skin [Skin] instance that will be applied to some table children. Defaults to [Scene2DSkin.defaultSkin].
@@ -389,10 +462,14 @@ inline fun <S> KWidget<S>.stack(
  * @return a [Table] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.table(
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KTableWidget.(S) -> Unit = {}
-): KTableWidget = actor(KTableWidget(skin), init)
+): KTableWidget {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KTableWidget(skin), init)
+}
 
 /**
  * @param text initial text displayed by the area. Defaults to empty string.
@@ -404,12 +481,16 @@ inline fun <S> KWidget<S>.table(
  * @return a [TextArea] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.textArea(
   text: String = "",
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl TextArea).(S) -> Unit = {}
-): TextArea = actor(TextArea(text, skin, style), init)
+): TextArea {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(TextArea(text, skin, style), init)
+}
 
 /**
  * @param text initial text displayed by the field. Defaults to empty string.
@@ -421,12 +502,16 @@ inline fun <S> KWidget<S>.textArea(
  * @return a [TextField] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.textField(
   text: String = "",
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl TextField).(S) -> Unit = {}
-): TextField = actor(TextField(text, skin, style), init)
+): TextField {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(TextField(text, skin, style), init)
+}
 
 /**
  * @param text will be displayed as [TextButton] text.
@@ -438,12 +523,16 @@ inline fun <S> KWidget<S>.textField(
  * @return a [TextButton] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.textButton(
   text: String,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KTextButton.(S) -> Unit = {}
-): KTextButton = actor(KTextButton(text, skin, style), init)
+): KTextButton {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KTextButton(text, skin, style), init)
+}
 
 /**
  * @param deadzone the distance from the center of the touchpad required for the knob to be moved.
@@ -455,12 +544,16 @@ inline fun <S> KWidget<S>.textButton(
  * @return a [Touchpad] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.touchpad(
   deadzone: Float,
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: (@Scene2dDsl Touchpad).(S) -> Unit = {}
-): Touchpad = actor(Touchpad(deadzone, skin, style), init)
+): Touchpad {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(Touchpad(deadzone, skin, style), init)
+}
 
 /**
  * @param style name of the widget style. Defaults to [defaultStyle].
@@ -471,11 +564,15 @@ inline fun <S> KWidget<S>.touchpad(
  * @return a [Tree] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.tree(
   style: String = defaultStyle,
   skin: Skin = Scene2DSkin.defaultSkin,
   init: KTreeWidget.(S) -> Unit = {}
-): KTreeWidget = actor(KTreeWidget(skin, style), init)
+): KTreeWidget {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KTreeWidget(skin, style), init)
+}
 
 /**
  * @param init will be invoked with the widget as "this". Consumes actor container (usually a [Cell] or [Node]) that
@@ -484,6 +581,10 @@ inline fun <S> KWidget<S>.tree(
  * @return a [VerticalGroup] instance added to this group.
  */
 @Scene2dDsl
+@OptIn(ExperimentalContracts::class)
 inline fun <S> KWidget<S>.verticalGroup(
   init: KVerticalGroup.(S) -> Unit = {}
-): KVerticalGroup = actor(KVerticalGroup(), init)
+): KVerticalGroup {
+  contract { callsInPlace(init, InvocationKind.EXACTLY_ONCE) }
+  return actor(KVerticalGroup(), init)
+}