+++ scene2d/src/test/kotlin/ktx/scene2d/tooltipTest.kt
@@ -25,32 +25,38 @@ class TooltipFactoriesTest : NeedsLibGDX() {
   @Test
   fun `should add TextTooltip with init block`() {
     val actor = Actor()
+    val variable: Int
 
     val tooltip = actor.textTooltip("Test.") {
       // Changing Label color:
       color = Color.BLUE
+      variable = 42
     }
 
     assertNotNull(tooltip)
     assertTrue(tooltip in actor.listeners)
     assertEquals("Test.", tooltip.actor.text.toString())
     assertEquals(Color.BLUE, tooltip.actor.color)
+    assertEquals(42, variable)
   }
 
   @Test
   fun `should add Tooltip with init block`() {
     val actor = Actor()
+    val variable: Int
 
     val tooltip = actor.tooltip {
       // Changing Table color:
       color = Color.BLUE
       // Adding child to Table content:
       label("Test.")
+      variable = 42
     }
 
     assertNotNull(tooltip)
     assertTrue(tooltip in actor.listeners)
     assertEquals("Test.", (tooltip.actor.children.first() as Label).text.toString())
     assertEquals(Color.BLUE, tooltip.actor.color)
+    assertEquals(42, variable)
   }
 }